chore: fix 17 pre-existing clippy errors across verdict queue + MCP + storage
Mechanical fixes: uninlined format args (10), unnecessary casts (4),
unwrap_or_default (1), too_many_arguments allow (1), expect allow+WHY (1).
Zero behavior changes.

+++ cynic-kernel/src/storage/reconnectable.rs
@@ -316,7 +316,7 @@ impl StoragePort for ReconnectableStorage {
     async fn queue_status_counts(&self) -> Result<(u32, u32, u32, u32), StorageError> {
         let (p, s, c, f) = self.current().queue_status_counts().await?;
         // Cast u64 to u32 for health endpoint compatibility (shouldn't overflow in practice)
-        Ok((p as u32, s as u32, c as u32, f as u32))
+        Ok((p, s, c, f))
     }
 }
